Trait dwt::Transform

pub trait Transform<T> {
    fn transform(&mut self, operation: Operation, wavelet: &Wavelet<T>, level: usize);
}

The transform.

Required Methods

fn transform(&mut self, operation: Operation, wavelet: &Wavelet<T>, level: usize)

Perform the transform.

The number of points should be divisible by 2^level. If the operation is forward, the data are replaced by the approximation and detail coefficients stored in the first and second halves of data, respectively. If the operation is inverse, the data are assumed to be stored according to the above convention.